[review Fix]  abstract the rate limiter into a separate file (#22643)
